[python3] robot: Made code directly Python 2.7/3.3+ compatible by applying (only necessary) compatible 2to3 diffs and using six for other cases.
--HG--
extra : transplant_source : %3D%A3%1C%14%60%28%E77/%B3g/%D9%1A%BB%A6%D4%28%93r

+++ src/robot/reporting/jsexecutionresult.py
@@ -35,7 +35,7 @@ class JsExecutionResult(object):
             'stats': statistics,
             'errors': errors,
             'baseMillis': basemillis,
-            'generatedMillis': long(time.mktime(gentime) * 1000) - basemillis,
+            'generatedMillis': int(time.mktime(gentime) * 1000) - basemillis,
             'generatedTimestamp': utils.format_time(gentime, gmtsep=' ')
         }
